(chore) move some helper files and hooks to component subfolders
