+++ frontend/src/components/SaturationFilter/hooks/useSaturationFilter.tsx
@@ -2,7 +2,7 @@ import { FilterInterface } from "components/list/types/list.types";
 import { colorMap } from "helpers/list.api";
 import React, { useEffect, useState } from "react";
 import { BiArrowToLeft, BiArrowToRight } from "react-icons/bi";
-import * as S from "./SaturationFilter.style";
+import * as S from "../SaturationFilter.style";
 
 export function useSaturationFilter(setFilter) {
 	const [saturationFilter, setSaturationFilter] = useState<
