+++ frontend/src/components/newlist/helpers/filterFalsyValues.ts
