+++ frontend/src/components/newlist/hooks/useNewList.tsx
@@ -3,10 +3,10 @@ import { newListState } from "components/newlist/state/newList.atom";
 import type { FocusIndex } from "components/newlist/types/newList.types";
 import { NewListFromClientInput } from "gql/codegen-output";
 import { useMutateCreateList } from "gql/hooks/list-mutate";
-import { filterFalsy } from "helpers/filterFalsyValues";
 import { useRouteProps } from "hooks/routerHooks";
 import { useCallback, useEffect, useMemo, useState } from "react";
 import { useRecoilState } from "recoil";
+import { filterFalsy } from "../helpers/filterFalsyValues";
 import NewListTerm from "../sub/NewListTerm";
 
 /*
