+++ frontend/src/helpers/list.api.ts
@@ -7,11 +7,11 @@ type Direction = "forwards" | "backwards";
  * @param {Object} listObj
  */
 export const extractSessionsByDirection = (list, direction: Direction) => {
-    // @todo: sessions currently don't exist as list.sessions entries.
-    const sessionsByDirection = list.sessions?.filter(
-        (sess) => sess.direction === direction
-    );
-    return [sessionsByDirection.length, sessionsByDirection];
+	// @todo: sessions currently don't exist as list.sessions entries.
+	const sessionsByDirection = list.sessions?.filter(
+		(sess) => sess.direction === direction
+	);
+	return [sessionsByDirection.length, sessionsByDirection];
 };
 
 /**
@@ -19,40 +19,24 @@ export const extractSessionsByDirection = (list, direction: Direction) => {
  * @param {Object} term
  */
 export const termSessionsByDirection = (term: Term, direction: Direction) => {
-    return term.history!.filter((session) => session?.direction === direction);
+	return term.history!.filter((session) => session?.direction === direction);
 };
 
 export const colorMap = {
-    0: "orangered",
-    1: "goldenrod",
-    2: "seagreen",
-    3: "deepskyblue",
-    4: "blueviolet",
+	0: "orangered",
+	1: "goldenrod",
+	2: "seagreen",
+	3: "deepskyblue",
+	4: "blueviolet",
 };
 
 export const colorBySaturation = (
-    saturation: { forwards: number; backwards: number } | number | undefined | string,
-    direction: Direction | undefined = undefined
+	saturation: { forwards: number; backwards: number } | number | undefined | string,
+	direction: Direction | undefined = undefined
 ) => {
-    if (typeof saturation === "number" || typeof saturation === "string") {
-        return colorMap[Number(saturation)];
-    } else if (direction && saturation) {
-        return colorMap[saturation[direction]];
-    }
+	if (typeof saturation === "number" || typeof saturation === "string") {
+		return colorMap[Number(saturation)];
+	} else if (direction && saturation) {
+		return colorMap[saturation[direction]];
+	}
 };
-
-/**
- * Takes a list, returns new list.state based on review session that was just completed.
- */
-export function maybeUpdateListStateAfterReview(list, direction: Direction) {
-    const [previousSessionLength] = extractSessionsByDirection(list, direction);
-
-    switch (previousSessionLength) {
-        case 0:
-            return { ...list.state, [direction]: "seeding" };
-        case 2:
-            return { ...list.state, [direction]: "seeded" };
-        default:
-            return list.state;
-    }
-}
